// src/types/REST/api/generated/models/AboutTeamMemberComponent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AboutTeamMemberComponentImage } from './AboutTeamMemberComponentImage';
import {
    AboutTeamMemberComponentImageFromJSON,
    AboutTeamMemberComponentImageFromJSONTyped,
    AboutTeamMemberComponentImageToJSON,
} from './AboutTeamMemberComponentImage';

/**
 * 
 * @export
 * @interface AboutTeamMemberComponent
 */
export interface AboutTeamMemberComponent {
    /**
     * 
     * @type {number}
     * @memberof AboutTeamMemberComponent
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof AboutTeamMemberComponent
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof AboutTeamMemberComponent
     */
    position?: string;
    /**
     * 
     * @type {string}
     * @memberof AboutTeamMemberComponent
     */
    linkedin?: string;
    /**
     * 
     * @type {string}
     * @memberof AboutTeamMemberComponent
     */
    twitter?: string;
    /**
     * 
     * @type {AboutTeamMemberComponentImage}
     * @memberof AboutTeamMemberComponent
     */
    image?: AboutTeamMemberComponentImage;
}

/**
 * Check if a given object implements the AboutTeamMemberComponent interface.
 */
export function instanceOfAboutTeamMemberComponent(value: object): value is AboutTeamMemberComponent {
    return true;
}

export function AboutTeamMemberComponentFromJSON(json: any): AboutTeamMemberComponent {
    return AboutTeamMemberComponentFromJSONTyped(json, false);
}

export function AboutTeamMemberComponentFromJSONTyped(json: any, ignoreDiscriminator: boolean): AboutTeamMemberComponent {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'position': json['position'] == null ? undefined : json['position'],
        'linkedin': json['linkedin'] == null ? undefined : json['linkedin'],
        'twitter': json['twitter'] == null ? undefined : json['twitter'],
        'image': json['image'] == null ? undefined : AboutTeamMemberComponentImageFromJSON(json['image']),
    };
}

export function AboutTeamMemberComponentToJSON(value?: AboutTeamMemberComponent | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'position': value['position'],
        'linkedin': value['linkedin'],
        'twitter': value['twitter'],
        'image': AboutTeamMemberComponentImageToJSON(value['image']),
    };
}